+++ src/black/mode.py
@@ -215,6 +215,7 @@ class Preview(Enum):
     remove_lone_list_item_parens = auto()
     pep646_typed_star_arg_type_var_tuple = auto()
     always_one_newline_after_import = auto()
+    generic_type_def_wrapping = auto()
 
 
 UNSTABLE_FEATURES: set[Preview] = {